Puts all overview panels' layouts and functions into their own files
